+++ src/Am.Keyward.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Am.Keyward.Core;
-
-public class Class1
-{
-
-}